feat(Main.vue): file upload & echo
Add a v-btn and make it to be able to fire click event to an other file input component when the v-btn being cliked.

Judge whether the uploaded file is an image.

Make URL from data in memory uploaded from file.

Show image with URL made above.

It works well!