fix: cross domain problem
Fix the cross domain problem by configuring the proxy in file vue.config.js.

It works.

P.S: Indeed the problem has been solved but it seems that the problem exists yet due to the project not being restart (this means the browser, the daemon & the vscode all should be restart after configuring).